+++ pkg/models/user.go
@@ -131,16 +131,16 @@ type Options struct {
 // UserStat is a helper struct to hold statistics about the whole app.
 type UserStat struct {
 	// PostCount is a number of posts of such user.
-	PostCount int `default:0`
+	PostCount int64 `default:0`
 
 	// ReactionCount tells the number of interactions (stars given).
-	ReactionCount int `default:0`
+	ReactionCount int64 `default:0`
 
 	// FlowerCount is basically a number of followers.
-	FlowerCount int `default:0`
+	FlowerCount int64 `default:0`
 
 	// ShadeCount is basically a number of blockers.
-	ShadeCount int `default:0`
+	ShadeCount int64 `default:0`
 
 	// Searched is a special boolean used by the search engine to mark who is to be shown in search results.
 	Searched bool `default:true`
